// frontend/src/gen/api/models/PostMapResponse.ts
export interface PostMapResponse {
  /**
   *
   * @type {Coordinate}
   * @memberof PostMapResponse
   */
  coodinate: Coordinate;
  /**
   *
   * @type {Date}
   * @memberof PostMapResponse
   */
  createdAt: Date;
  /**
   *
   * @type {ProblemMapResponse}
   * @memberof PostMapResponse
   */
  descriptions: ProblemMapResponse | null;
  /**
   *
   * @type {string}
   * @memberof PostMapResponse
   */
  id: string;
  /**
   *
   * @type {boolean}
   * @memberof PostMapResponse
   */
  isSolved: boolean;
  /**
   *
   * @type {ProblemMapResponse}
   * @memberof PostMapResponse
   */
  photoField: ProblemMapResponse | null;
  /**
   *
   * @type {ProblemForPost}
   * @memberof PostMapResponse
   */
  problem: ProblemForPost;
  /**
   *
   * @type {Date}
   * @memberof PostMapResponse
   */
  updatedAt: Date | null;
  /**
   *
   * @type {UpdatedBy}
   * @memberof PostMapResponse
   */
  updatedBy: UpdatedBy | null;
  /**
   *
   * @type {UserForPost}
   * @memberof PostMapResponse
   */
  user: UserForPost | null;
}